tests/wpt/tests/navigation-api/navigate-event/cross-window/click-samedocument-crossorigin.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body>

<script>
async_test(t => {
  const url = new URL("resources/cross-origin-iframe-helper.html", location.href);
  url.hostname = get_host_info().REMOTE_HOST;
  const iframe = document.createElement("iframe")
  iframe.src = url;
  iframe.name = "windowname";
  document.body.append(iframe);

  url.hash = "#foo";
  const link = document.createElement("a");
  link.href = url;
  link.target = iframe.name;
  document.body.append(link);

  window.onmessage = t.step_func_done(e => {
    assert_equals(e.data.navigationType, "push", "navigationType");
    assert_true(e.data.cancelable, "cancelable");
    assert_true(e.data.canIntercept, "canIntercept");
    assert_false(e.data.userInitiated, "userInitiated");
    assert_true(e.data.hashChange, "hashChange");
    assert_equals(e.data.formData, null, "formData");
    assert_equals(e.data.destination.url, link.href, "destination.url");
    assert_true(e.data.destination.sameDocument, "destination.sameDocument");
    assert_equals(e.data.destination.key, "", "destination.key");
    assert_equals(e.data.destination.id, "", "destination.id");
    assert_equals(e.data.destination.index, -1, "destination.index");
  });

  navigation.onnavigate = t.unreached_func("onnavigate must not fire in the source window");
  window.onload = t.step_func(() => link.click());
}, "clicking on an <a> element that navigates same-document targeting a cross-origin window");
</script>
